fix: check delegatecall
In case the hyvm is in a context of call, a check
needs to be done to verify if a delegatecall is executed.
There could be a case where it would delegate to a selfdestruct.
With this fix, no delegatecall is possible in a call context.

Note: it seems __codesize(CONSTRUCTOR) does not take into
account all opcodes in constructor